workWave_server/requirements.txt
# psycopg2==2.9.7
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.34.0
cryptography==41.0.3
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.2.5
django-appconf==1.0.5
django-cors-headers==4.2.0
django-cryptography==1.1
django-debug-toolbar==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
